feat: 결제 내역 데이터 가져오기 2차
* 상위 컴포넌트에서 API 요청하도록 수정
