+++ front-end/src/App.css
@@ -1,9 +1,22 @@
+.app {
+  display: flex;
+  flex-direction: column;
+  height: 100vh;
+}
+
 .app-header {
   margin: 10vh auto 0;
   width: 80vw;
-  height: 10vh;
 }
 .app-header .abs-full-width-button .dropdown-toggle,
 .app-header .abs-full-width-button .dropdown-menu {
   width: 100%;
+}
+
+.app-main {
+  margin: 0 auto;
+  width: 90vw;
+  flex: 1 0 auto;
+  overflow: auto;
+  position: relative;
 }
