+++ front-end/src/App.js
@@ -1,7 +1,9 @@
 import './App.css';
 import {Dropdown, DropdownButton} from 'react-bootstrap';
-import {useState} from 'react';
+import {useEffect, useReducer, useState} from 'react';
+import DailyPaymentContainer from './components/DailyPaymentContainer';
 import SummaryBySign from './components/SummaryBySign';
+import * as Api from './common/Api';
 
 function App() {
   const [currentMonth, setCurrentMonth] = useState(new Date().getMonth());
@@ -17,6 +19,25 @@ function App() {
         {i + 1}월
       </Dropdown.Item>);
 
+  const [state, dispatch] = useReducer((_, action) => ({
+    processing: action.processing,
+    payments: action.data ?? []
+  }), { processing: false, payments: [] });
+
+  useEffect(() => {
+    fetchData().then();
+
+    async function fetchData() {
+      const monthDateStr = ``+
+        `${new Date().getFullYear()}-${String(currentMonth + 1).padStart(2, '0')}`;
+
+      await Api.getWithDispatch(dispatch, '/payments', {
+        start: monthDateStr,
+        end: monthDateStr
+      });
+    }
+  }, [currentMonth]);
+
   return (
     <div className="app">
       <header className="app-header">
@@ -28,10 +49,13 @@ function App() {
           {monthDropdownItems}
         </DropdownButton>
         <SummaryBySign
-          currentMonth={currentMonth}
+          payments={state.payments}
+          processing={state.processing}
           mt="1"
         />
       </header>
+      <main className="app-main">
+      </main>
     </div>
   );
 }
